+++ prosodic/langs/__init__.py
@@ -1,1 +1,2 @@
+from .. import *
 from .langs import *
