+++ prosodic/langs/english/__init__.py
@@ -1,1 +1,2 @@
+from .. import *
 from .english import *
