+++ prosodic/parsing/__init__.py
@@ -1,3 +1,4 @@
+from .. import *
 from .constraints import *
 from .meter import *
 from .parses import *
