+++ prosodic/sents/__init__.py
@@ -1,3 +1,4 @@
+from .. import *
 from .sents import *
 from .trees import *
 from .syntax import *
