+++ prosodic/texts/__init__.py
@@ -1,3 +1,4 @@
+from .. import *
 from .texts import *
 from .stanzas import *
 from .lines import *
