+++ prosodic/texts/texts.py
@@ -2,12 +2,14 @@ from ..imports import *
 from ..words import WordTokenList, WordToken
 
 NUMBUILT = 0
-DEFAULT_COMBINE_BY = 'line'
+DEFAULT_COMBINE_BY = "line"
 
-class TextModel(WordTokenList):
+
+class TextModel(Entity):
     """
     A class that represents a text structure, comprised of WordTokens.
     """
+
     is_text = True
     prefix = "text"
 
@@ -42,7 +44,9 @@ class TextModel(WordTokenList):
         Returns:
             None
         """
-        if isinstance(children,str):
+        global OBJECTS
+
+        if isinstance(children, str):
             txt = children
             children = []
 
@@ -50,74 +54,58 @@ class TextModel(WordTokenList):
             raise ValueError(
                 "must provide either txt string or filename or token dataframe"
             )
-        self._text = None
-
-        #log.debug("Cleaning and getting text")
-        txt = clean_text(get_txt(txt, fn)).strip()
-
-        #log.debug(f"Cleaned text: {txt[:100]}...")
-
-        self._txt = txt
-        self._fn = fn
-
-        #log.debug(f"Setting language: {lang}")
-
-        self.lang = lang if lang else detect_lang(txt)
-        #log.debug(f"Language set to: {self.lang}")
-
-        was_quiet = logmap.quiet
-        #log.debug(f"Was quiet: {was_quiet}")
-
+        self._txt = clean_text(get_txt(txt, fn)).strip()
+        self.lang = lang if lang else detect_lang(self._txt)
+        self._key = f"{self.nice_type_name}({self.hash})"
         if not children:
-            #log.debug("No children provided, processing text")
-
-            numwords = len(txt.split())
-            #log.debug(f"Number of words: {numwords}")
-
-            if was_quiet and numwords > 1000:
-                #log.debug("Setting logmap to not quiet due to large text")
-                logmap.quiet = False
-
-            #log.debug(f"building text with {numwords:,} words")
             if tokens_df is None:
-                #log.debug("Tokenizing text into DataFrame")
                 tokens_df = tokenize_sentwords_df(txt)
-            children = [
-                WordToken(lang=self.lang, parent=self, text=self, **row.to_dict())
-                for _, row in progress_bar(
-                    list(tokens_df.iterrows()),
-                    progress=len(tokens_df)>=1000, 
-                    desc='Building long text'
-                )
-            ]
-        #log.debug("Initializing parent class")
-        super().__init__(children=children, parent=parent, **kwargs)
 
-    
+            children = WordTokenList(parent=self)
+            for _, row in progress_bar(
+                list(tokens_df.iterrows()),
+                progress=len(tokens_df) >= 1000,
+                desc="Building long text",
+            ):
+                children.append(
+                    WordToken(
+                        lang=self.lang,
+                        parent=children,
+                        text=self,
+                        **row.to_dict(),
+                    )
+                )
+        super().__init__(children=children, txt=txt, key=self._key, parent=parent, text=self, num=0, **kwargs)
+        # OBJECTS[self.key] = self
 
     @cached_property
     def stanzas(self):
         from ..texts.stanzas import StanzaList
+
         return StanzaList.from_wordtokens(self, text=self)
-    
+
     @cached_property
     def lines(self):
         from ..texts.lines import LineList
+
         return LineList.from_wordtokens(self, text=self)
-    
+
     @cached_property
     def lineparts(self):
         from ..texts.lines import LinePartList
+
         return LinePartList.from_wordtokens(self, text=self)
-    
+
     @cached_property
     def sents(self):
         from ..sents.sents import SentenceList
+
         return SentenceList.from_wordtokens(self, text=self)
-    
+
     @cached_property
     def sentparts(self):
         from ..sents.sents import SentPartList
+
         return SentPartList.from_wordtokens(self, text=self)
 
     def to_hash(self) -> str:
@@ -131,15 +119,24 @@ class TextModel(WordTokenList):
 
     @property
     def hash(self):
-        return encode_hash(serialize({'txt':self.txt, 'lang':self.lang}))
+        return encode_hash(serialize({"txt": self._txt, "lang": self.lang}))
 
-    @cached_property
+    @property
     def key(self):
-        return f'{self.nice_type_name}({self.hash})'
-
+        if self._key is None:
+            self._key = f"{self.nice_type_name}({self.hash})"
+        return self._key
 
     # @stash.stashed_result
-    def parse(self, combine_by: Literal['line','sent'] = DEFAULT_COMBINE_BY, num_proc=None, lim=None, force=False, meter=None, **meter_kwargs):
+    def parse(
+        self,
+        combine_by: Literal["line", "sent"] = DEFAULT_COMBINE_BY,
+        num_proc=None,
+        lim=None,
+        force=False,
+        meter=None,
+        **meter_kwargs,
+    ):
         """
         Parse the text.
 
@@ -150,6 +147,7 @@ class TextModel(WordTokenList):
             Any: The parsed result.
         """
         from ..parsing.parselists import ParseListList
+
         self._parses = ParseListList(
             self.parse_iter(
                 combine_by=combine_by,
@@ -159,19 +157,30 @@ class TextModel(WordTokenList):
                 lim=lim,
                 **meter_kwargs,
             ),
-            parent=self
+            parent=self,
         )
         return self._parses
 
-    def parse_iter(self, combine_by: Literal['line','sent'] = DEFAULT_COMBINE_BY, num_proc=None, lim=None, force=False, meter=None, **meter_kwargs):
+    def parse_iter(
+        self,
+        combine_by: Literal["line", "sent"] = DEFAULT_COMBINE_BY,
+        num_proc=None,
+        lim=None,
+        force=False,
+        meter=None,
+        **meter_kwargs,
+    ):
         from ..parsing.parselists import ParseList
+
         meter = self.get_meter(meter=meter, **meter_kwargs)
         if combine_by and meter.parse_unit == combine_by:
             combine_by = None
-        
+
         last_unit = None
-        units=[]
-        for parse_list in meter.parse_text_iter(self, num_proc=num_proc, force=force, lim=lim):
+        units = []
+        for parse_list in meter.parse_text_iter(
+            self, num_proc=num_proc, force=force, lim=lim
+        ):
             parsed_ent = self.match(parse_list.parent)
             parse_list.parent = parsed_ent
             parsed_ent._parses = parse_list
@@ -183,7 +192,7 @@ class TextModel(WordTokenList):
                     new_parselist = ParseList.from_combinations(units, parent=this_unit)
                     this_unit._parses = new_parselist
                     yield new_parselist
-                    units=[]
+                    units = []
                 units.append(parse_list)
                 last_unit = this_unit
 
@@ -191,9 +200,7 @@ class TextModel(WordTokenList):
             new_parselist = ParseList.from_combinations(units, parent=this_unit)
             this_unit._parses = new_parselist
             yield new_parselist
-        
 
-        
 
 @stash.stashed_result
 def Text(
@@ -204,7 +211,9 @@ def Text(
     children: Optional[list] = [],
     tokens_df: Optional[pd.DataFrame] = None,
 ):
-    return TextModel(txt=txt, fn=fn, lang=lang, parent=parent, children=children, tokens_df=tokens_df)
+    return TextModel(
+        txt=txt, fn=fn, lang=lang, parent=parent, children=children, tokens_df=tokens_df
+    )
 
 
 class TextList(EntityList):
