+++ prosodic/words/__init__.py
@@ -1,6 +1,8 @@
 from .. import *
 from .tokenizers import *
-from .words import *
-from .syllables import *
 from .phonemes import *
+from .syllables import *
+from .wordform import *
+from .wordtype import *
+from .wordtoken import *
 from .wordtokenlist import *
